[Rule][Test][WinConditions]
I have modelled the win conditions as three different predicates.

One for when all the X coordinates are the same

One for when all the Y coordinates are the same

One for when the delta of X and Y are the same, which means a diagonal 
is taken place.


In order to achieve the last (third) predicate the coordinate system was 
modeled symetric, meaning the middle coordinate is  centered at (0,0).


Finnally there are is the win rule, which applies these three predicates 
onto the moves made by a given player, in order to determine if it 
complies with any of the three predicates given below, which are the win 
conditions for the game.